requirements.txt
# NLP and Text Processing
sentence-transformers>=2.2.2
transformers>=4.36.0
spacy>=3.7.0
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl
nltk>=3.8.1
regex>=2023.10.3

# Vector Database & Embeddings
pinecone-client>=2.2.4
chromadb>=0.4.15
faiss-cpu>=1.7.4

# LLM Integration
openai>=1.3.0
llama-cpp-python>=0.2.20
huggingface-hub>=0.19.0

# Web Framework
streamlit>=1.28.0
fastapi>=0.95.2
uvicorn>=0.22.0

# Data Processing
pandas>=2.0.3
numpy>=1.24.3
tqdm>=4.65.0

# Utility Libraries
python-dotenv>=1.0.0
loguru>=0.7.0
typer>=0.9.0
pydantic>=2.3.0

# Development & Testing
pytest>=7.4.0
black>=23.7.0
mypy>=1.5.1
isort>=5.12.0